services/ml-pipeline/requirements.txt
# ML Pipeline Dependencies

# Core ML
xgboost==2.0.3
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# gRPC for inference server
grpcio==1.60.0
grpcio-tools==1.60.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Scheduling
APScheduler==3.10.4

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
